+++ js/get-rand-arr-el.js
@@ -1,0 +1,4 @@
+import {getRandomInteger} from './get-rand-int.js';
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomArrayElement};
