+++ js/get-uniq-rand-int-from-range.js
@@ -1,0 +1,19 @@
+import {getRandomInteger} from './get-rand-int.js';
+function getUniqRandomIntFromRange (min, max) {
+  const previousValues = [];
+
+  return function () {
+    let currentValue = getRandomInteger(min, max);
+    if (previousValues.length >= (max - min + 1)) {
+      console.error('Перебраны все числа из диапазона от ' + min + ' до ' + max);
+      return null;
+    }
+    while (previousValues.includes(currentValue)) {
+      currentValue = getRandomInteger(min, max);
+    }
+    previousValues.push(currentValue);
+    return currentValue;
+  };
+}
+
+export {getUniqRandomIntFromRange};
